+++ icd-db/src/main/scala/csw/services/icd/db/IcdDb.scala
@@ -1,5 +1,8 @@
 package csw.services.icd.db
 
+import com.fasterxml.jackson.core.util.DefaultPrettyPrinter
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
+
 import java.io.File
 import com.typesafe.config.{Config, ConfigFactory}
 import csw.services.icd.*
@@ -11,6 +14,7 @@ import csw.services.icd.fits.IcdFits
 import diffson.playJson.DiffsonProtocol
 import icd.web.shared.IcdModels.{EventModel, ImageModel, ReceiveCommandModel}
 import icd.web.shared.{BuildInfo, HtmlHeadings, IcdModels, PdfOptions, SubsystemWithVersion}
+import io.swagger.util.Yaml
 import io.swagger.v3.parser.OpenAPIV3Parser
 import io.swagger.v3.parser.core.models.ParseOptions
 import io.swagger.v3.parser.util.DeserializationUtils
@@ -159,6 +163,10 @@ object IcdDb extends App {
       c.copy(archived = Some(x))
     } text "Generates an 'Archived Items' report for all subsystems (or the given one) to the given file in a format based on the file's suffix (html, pdf, csv)"
 
+    opt[File]("alarms") valueName "<outputFile>" action { (x, c) =>
+      c.copy(alarms = Some(x))
+    } text "Generates an 'Alarms' report for all subsystems (or the given one) to the given file in a format based on the file's suffix (html, pdf, csv)"
+
     opt[Unit]("allSubsystems") action { (_, c) =>
       c.copy(allSubsystems = Some(()))
     } text "Include all subsystems in searches for publishers, subscribers, etc. while generating API doc (Default: only consider the one subsystem)"
@@ -254,6 +262,7 @@ object IcdDb extends App {
     options.diff.foreach(diffVersions)
     options.missing.foreach(missingItemsReport)
     options.archived.foreach(file => archivedItemsReport(file, Some(pdfOptions)))
+    options.alarms.foreach(file => alarmsReport(file, Some(pdfOptions)))
     options.listData.foreach(s => listData(db, s))
     options.allUnits.foreach(_ => printAllUsedUnits(db))
 
@@ -390,6 +399,14 @@ object IcdDb extends App {
       ArchivedItemsReport(db, maybeSv, maybePdfOptions, new HtmlHeadings).saveToFile(file, pdfOptions)
     }
 
+    // --alarms option
+    def alarmsReport(file: File, maybePdfOptions: Option[PdfOptions]): Unit = {
+      val maybeSv = options.subsystem
+        .map(SubsystemAndVersion(_))
+        .map(s => SubsystemWithVersion(s.subsystem, s.maybeVersion, options.component))
+      AlarmsReport(db, maybeSv, maybePdfOptions, new HtmlHeadings).saveToFile(file, pdfOptions)
+    }
+
     // --generate code in the given file
     def generate(file: File): Unit = {
       file.getName.split("\\.").last match {
@@ -446,8 +463,8 @@ case class IcdDb(
     port: Int = IcdDbDefaults.defaultPort
 ) {
   // Cleanup databases from earlier releases
-  IcdDbDefaults.deleteDatabase(host, port, "icds")
-  IcdDbDefaults.deleteDatabase(host, port, "icds2")
+//  IcdDbDefaults.deleteDatabase(host, port, "icds")
+//  IcdDbDefaults.deleteDatabase(host, port, "icds2")
   IcdDbDefaults.deleteDatabase(host, port, "icds3")
 
   val db: DB    = IcdDbDefaults.connectToDatabase(host, port, dbName)
@@ -493,7 +510,7 @@ case class IcdDb(
 
   // Check for obvious errors, such as duplicate event or parameter names after ingesting model files for subsystem
   private def checkPostIngest(subsystem: String): List[Problem] = {
-    // Check for min > max in parameter fields and return list of error messages for the parameter list
+    // Check for min > max and defaultValue in parameter fields and return list of error messages for the parameter list
     def getParamMinMaxProblems(parameterList: List[IcdModels.ParameterModel]): List[String] = {
       def strToDouble(s: String): Double = {
         s match {
@@ -502,6 +519,71 @@ case class IcdDb(
           case _               => s.toDoubleOption.getOrElse(0.0)
         }
       }
+
+      // Check that default parameter value is valid for declared type
+      def checkDefaultParamValue(p: IcdModels.ParameterModel): Option[String] = {
+        if (p.defaultValue.isEmpty)
+          None
+        else {
+          if (p.maybeEnum.isDefined) {
+            if (p.maybeEnum.get.contains(p.defaultValue))
+              None
+            else {
+              val choices = p.maybeEnum.get.mkString(", ")
+              Some(s"In parameter ${p.name}, defaultValue ${p.defaultValue} is invalid (Should be one of: $choices)")
+            }
+          }
+          else if (p.maybeType.isDefined) {
+            p.maybeType.get match {
+              case "boolean" =>
+                if (p.defaultValue.toBooleanOption.nonEmpty)
+                  None
+                else
+                  Some(s"In parameter ${p.name}, defaultValue ${p.defaultValue} is invalid (Should be one of: true, false)")
+              case "integer" =>
+                None
+                if (p.defaultValue.toIntOption.nonEmpty)
+                  None
+                else
+                  Some(s"In parameter ${p.name}, defaultValue ${p.defaultValue} is invalid (Should be an integer value)")
+              case "byte" =>
+                None
+                if (p.defaultValue.toByteOption.nonEmpty)
+                  None
+                else
+                  Some(s"In parameter ${p.name}, defaultValue ${p.defaultValue} is invalid (Should be a byte value)")
+              case "short" =>
+                None
+                if (p.defaultValue.toShortOption.nonEmpty)
+                  None
+                else
+                  Some(s"In parameter ${p.name}, defaultValue ${p.defaultValue} is invalid (Should be a short value)")
+              case "long" =>
+                None
+                if (p.defaultValue.toLongOption.nonEmpty)
+                  None
+                else
+                  Some(s"In parameter ${p.name}, defaultValue ${p.defaultValue} is invalid (Should be a long value)")
+              case "float" =>
+                None
+                if (p.defaultValue.toFloatOption.nonEmpty)
+                  None
+                else
+                  Some(s"In parameter ${p.name}, defaultValue ${p.defaultValue} is invalid (Should be a float value)")
+              case "double" =>
+                None
+                if (p.defaultValue.toDoubleOption.nonEmpty)
+                  None
+                else
+                  Some(s"In parameter ${p.name}, defaultValue ${p.defaultValue} is invalid (Should be a double value)")
+              case _ => None
+            }
+          }
+          else // if (p.maybeArrayType.isDefined) {
+            None
+        }
+      }
+
       val list = parameterList.flatMap { p =>
         val a =
           if (
@@ -518,7 +600,8 @@ case class IcdDb(
           if (p.maxItems.isDefined && p.minItems.isDefined && p.maxItems.get < p.minItems.get)
             Some(s"In parameter ${p.name}, maxItems (${p.maxItems.get}) < minItems (${p.minItems.get})")
           else None
-        List(a, b, c)
+        val d = checkDefaultParamValue(p)
+        List(a, b, c, d)
       }
       list.flatten
     }
@@ -604,12 +687,12 @@ case class IcdDb(
         } yield {
           val keyNameProblems =
             if (allowedFitsKeyNames.contains(k.name)) None
-            else Some(Problem("error", s"${prefix}.${event.name}: ${k.name} is not in the FITS dictionary"))
+            else Some(Problem("error", s"$prefix.${event.name}: ${k.name} is not in the FITS dictionary"))
           val channelNameProblems =
             if (k.channel.isEmpty || allowedChannels.contains(k.channel.get)) None
             else {
               val msg1 =
-                s"Event: ${prefix}.${event.name}, parameter: ${p.name}, FITS keyword: ${k.name}: Channel ${k.channel.get} is not defined.  "
+                s"Event: $prefix.${event.name}, parameter: ${p.name}, FITS keyword: ${k.name}: Channel ${k.channel.get} is not defined.  "
               val msg2 = if (allowedChannels.nonEmpty) s"Should be one of: ${allowedChannels.mkString(", ")}" else ""
               Some(Problem("error", msg1 + msg2))
             }
@@ -786,7 +869,7 @@ case class IcdDb(
   private def ingestConfig(stdConfig: StdConfig): List[Problem] = {
     // Ingest a single OpenApi file
     def ingestOpenApiFile(collectionName: String, fileName: String): List[Problem] = {
-      import scala.jdk.CollectionConverters._
+      import scala.jdk.CollectionConverters.*
       val parseOptions = new ParseOptions()
       parseOptions.setResolve(true)
       parseOptions.setResolveFully(true)
@@ -799,7 +882,11 @@ case class IcdDb(
         // Convert YAML to JSON if needed
         val jsonStr =
           if (fileName.endsWith(".yml") || fileName.endsWith(".yaml")) {
-            val yaml = io.swagger.util.Yaml.pretty().writeValueAsString(openAPI)
+            val mapper = Yaml.mapper()
+            // See https://stackoverflow.com/questions/67874510/spring-boot-2-5-0-and-invaliddefinitionexception-java-8-date-time-type-java-ti
+            mapper.registerModule(new JavaTimeModule())
+            val writer = mapper.writer(new DefaultPrettyPrinter())
+            val yaml   = writer.writeValueAsString(openAPI)
             DeserializationUtils.deserializeIntoTree(yaml, fileName).toPrettyString
           }
           else {
@@ -830,7 +917,9 @@ case class IcdDb(
       ingestOpenApiFiles(coll)
     }
     catch {
-      case t: Throwable => List(Problem("error", s"Internal error: $t"))
+      case t: Throwable =>
+        t.printStackTrace()
+        List(Problem("error", s"Internal error: $t"))
     }
   }
 
@@ -843,7 +932,7 @@ case class IcdDb(
    */
   //noinspection SameParameterValue
   private def ingestConfig(name: String, tmpName: String, config: Config): Unit = {
-    import play.api.libs.json.Reads._
+    import play.api.libs.json.Reads.*
 
     val jsObj = Json.parse(IcdValidator.toJson(config)).as[JsObject]