+++ icd-web-client/src/main/scala/icd/web/client/IcdWebClient.scala
@@ -69,6 +69,14 @@ case class IcdWebClient(csrfToken: String, inputDirSupported: Boolean) {
     showDetailButtons = false
   )
 
+  private val alarmsItem = NavbarPdfItem(
+    "Alarms",
+    "Generate and display an 'Alarms' report for the selected subsystem/component (or all subsystems)",
+    makeAlarmsReport,
+    showDocumentNumber = false,
+    showDetailButtons = false
+  )
+
   private val missingItem = NavbarPdfItem(
     "Missing",
     "Generate and display a 'Missing Items' report for the selected subsystems/components (or all subsystems)",
@@ -166,6 +174,7 @@ case class IcdWebClient(csrfToken: String, inputDirSupported: Boolean) {
     navbar.addItem(generateItem)
     navbar.addItem(graphItem)
     navbar.addItem(archiveItem)
+    navbar.addItem(alarmsItem)
     navbar.addItem(missingItem)
     navbar.addItem(fitsDictionaryItem)
     navbar.addItem(publishItem)
@@ -238,7 +247,7 @@ case class IcdWebClient(csrfToken: String, inputDirSupported: Boolean) {
   private def showPasswordDialog()(): Unit = {
     setSidebarVisible(false)
     setNavbarVisible(false)
-    val title = "TMT Interface Database System"
+    val title = "TIO Software Interface Database System"
     mainContent.setContent(passwordDialog, s"$title ${BuildInfo.version}")
   }
 
@@ -246,7 +255,7 @@ case class IcdWebClient(csrfToken: String, inputDirSupported: Boolean) {
   private def showStatus(maybeSubsystem: Option[String] = None, saveHistory: Boolean = true)(): Unit = {
     setSidebarVisible(false)
     currentView = StatusView
-    val title = "TMT Interface Database System"
+    val title = "TIO Software Interface Database System"
     if (saveHistory) {
       mainContent.setContent(statusDialog, s"$title ${BuildInfo.version}")
       pushState(viewType = StatusView, maybeSourceSubsystem = maybeSubsystem.map(SubsystemWithVersion(_, None, None)))
@@ -536,7 +545,7 @@ case class IcdWebClient(csrfToken: String, inputDirSupported: Boolean) {
 
   private def addLinkHandlers(): Unit = {
     val links = document.getElementsByTagName("a")
-    links.toArray.foreach(x => x.addEventListener("click", linkListener))
+    links.toArray.foreach(x => x.addEventListener("click", linkListener(_)))
   }
 
   // Called when the "History" item is selected
@@ -675,6 +684,21 @@ case class IcdWebClient(csrfToken: String, inputDirSupported: Boolean) {
         ClientRoutes.archivedItemsReport(maybeSv.get, options)
       else
         ClientRoutes.archivedItemsReportFull(options)
+    dom.window.open(uri) // opens in new window or tab
+  }
+
+  // Gets a PDF with an Alarms report for the currently selected subsystem API
+  private def makeAlarmsReport(options: PdfOptions): Unit = {
+    val maybeSv =
+      if (currentView == StatusView)
+        statusDialog.getSubsystemWithVersion
+      else selectDialog.subsystem.getSubsystemWithVersion()
+
+    val uri =
+      if (maybeSv.isDefined)
+        ClientRoutes.alarmsReport(maybeSv.get, options)
+      else
+        ClientRoutes.alarmsReportFull(options)
     dom.window.open(uri) // opens in new window or tab
   }