+++ icd-web-server/app/controllers/Application.scala
@@ -425,6 +425,69 @@ class Application @Inject() (
     }
 
   /**
+   * Returns the alarms report (PDF) for the given subsystem API
+   *
+   * @param subsystem        the source subsystem
+   * @param maybeVersion     the source subsystem's version (default: current)
+   * @param maybeComponent   optional component (default: all in subsystem)
+   * @param maybeOrientation If set, should be "portrait" or "landscape" (default: landscape)
+   * @param maybeFontSize    base font size for body text (default: 10)
+   */
+  def alarmsReport(
+      subsystem: String,
+      maybeVersion: Option[String],
+      maybeComponent: Option[String],
+      maybeOrientation: Option[String],
+      maybeFontSize: Option[Int],
+      maybeLineHeight: Option[String],
+      maybePaperSize: Option[String]
+  ) =
+    authAction.async {
+      val resp: Future[Option[Array[Byte]]] = appActor ? (
+        GetAlarmsReport(
+          subsystem,
+          maybeVersion,
+          maybeComponent,
+          PdfOptions(maybeOrientation, maybeFontSize, maybeLineHeight, maybePaperSize),
+          _
+        )
+      )
+      resp.map {
+        case Some(bytes) =>
+          Ok(bytes).as("application/pdf")
+        case None =>
+          NotFound
+      }
+    }
+
+  /**
+   * Returns the alarms report (PDF) for all current subsystems
+   *
+   * @param maybeOrientation If set, should be "portrait" or "landscape" (default: landscape)
+   * @param maybeFontSize    base font size for body text (default: 10)
+   */
+  def alarmsReportFull(
+      maybeOrientation: Option[String],
+      maybeFontSize: Option[Int],
+      maybeLineHeight: Option[String],
+      maybePaperSize: Option[String]
+  ) =
+    authAction.async {
+      val resp: Future[Option[Array[Byte]]] = appActor ? (
+        GetAlarmsReportFull(
+          PdfOptions(maybeOrientation, maybeFontSize, maybeLineHeight, maybePaperSize),
+          _
+        )
+      )
+      resp.map {
+        case Some(bytes) =>
+          Ok(bytes).as("application/pdf")
+        case None =>
+          NotFound
+      }
+    }
+
+  /**
    * Returns a missing items report (PDF) for the given subsystem/component API
    *
    * @param subsystem      the source subsystem
@@ -530,6 +593,7 @@ class Application @Inject() (
       maybeCommandLabels: Option[Boolean],
       maybeEventLabels: Option[Boolean],
       maybeGroupSubsystems: Option[Boolean],
+      maybeOnlySubsystems: Option[Boolean],
       maybeLayout: Option[String],
       maybeOverlap: Option[String],
       maybeSplines: Option[Boolean],
@@ -554,6 +618,7 @@ class Application @Inject() (
             commandLabels = maybeCommandLabels.getOrElse(defaultCommandLabels),
             eventLabels = maybeEventLabels.getOrElse(defaultEventLabels),
             groupSubsystems = maybeGroupSubsystems.getOrElse(defaultGroupSubsystems),
+            onlySubsystems = maybeOnlySubsystems.getOrElse(defaultOnlySubsystems),
             layout = maybeLayout.getOrElse(defaultLayout),
             overlap = maybeOverlap.getOrElse(defaultOverlap),
             splines = maybeSplines.getOrElse(defaultUseSplines),